Join the pause menu icon into the same icon row as the title screen
PauseScreenMixin was placing the icon into a hardcoded GridLayout cell
(row 3, col 0) that no longer lines up with the actual icon row in
26.2's pause menu layout, leaving it floating next to the Options
button instead. Switched to the same dynamic row-discovery approach
already used (and proven working) for the title screen, extracted into
a shared IconRowUtils helper used by both mixins.
