+++ src/main/java/dev/evvie/waylandcraft/mixin/PauseScreenMixin.java
@@ -1,18 +1,19 @@
 package dev.evvie.waylandcraft.mixin;
 
+import java.util.List;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.llamalad7.mixinextras.sugar.Local;
-
 import dev.evvie.waylandcraft.WaylandCraft;
 import dev.evvie.waylandcraft.WaylandCraftCommon;
+import dev.evvie.waylandcraft.gui.IconRowUtils;
 import dev.evvie.waylandcraft.gui.WaylandCraftSettingsScreen;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.components.SpriteIconButton;
-import net.minecraft.client.gui.layouts.GridLayout;
 import net.minecraft.client.gui.screens.PauseScreen;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
@@ -20,26 +21,29 @@ import net.minecraft.resources.Identifier;
 
 @Mixin(PauseScreen.class)
 public class PauseScreenMixin extends Screen {
-	
+
 	protected PauseScreenMixin() {
 		super(null);
 	}
-	
-	private SpriteIconButton button;
-	
-	@Inject(method = "createPauseMenu", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/layouts/GridLayout;arrangeElements()V"))
-	public void addButton(CallbackInfo info, @Local GridLayout layout) {
-		button = SpriteIconButton
+
+	@Inject(method = "createPauseMenu", at = @At("TAIL"))
+	public void addWaylandCraftButton(CallbackInfo info) {
+		List<AbstractWidget> row = IconRowUtils.findIconRow(this.children());
+
+		SpriteIconButton button = SpriteIconButton
 				.builder(Component.literal("waylandcraft"), (_) -> {Minecraft.getInstance().gui.setScreen(new WaylandCraftSettingsScreen(WaylandCraft.instance));}, true)
 				.sprite(Identifier.fromNamespaceAndPath(WaylandCraftCommon.MOD_ID, "logo"), 16, 16)
-				.width(20)
+				.width(IconRowUtils.ICON_ROW_BUTTON_WIDTH)
 				.build();
-		layout.addChild(button, 3, 0);
+
+		if(row.isEmpty()) {
+			button.setPosition(width / 2 - 124, height / 4 + 72 + 6);
+			this.addRenderableWidget(button);
+			return;
+		}
+
+		IconRowUtils.joinRow(row, button);
+		this.addRenderableWidget(button);
 	}
-	
-	@Inject(method = "createPauseMenu", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/layouts/GridLayout;visitWidgets(Ljava/util/function/Consumer;)V"))
-	public void offsetButton(CallbackInfo info) {
-		button.setX(button.getX() - 24);
-	}
-	
+
 }
