+++ src/main/java/dev/evvie/waylandcraft/mixin/TitleScreenMixin.java
@@ -1,9 +1,6 @@
 package dev.evvie.waylandcraft.mixin;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -12,6 +9,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import dev.evvie.waylandcraft.WaylandCraft;
 import dev.evvie.waylandcraft.WaylandCraftCommon;
+import dev.evvie.waylandcraft.gui.IconRowUtils;
 import dev.evvie.waylandcraft.gui.WaylandCraftSettingsScreen;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.AbstractWidget;
@@ -24,20 +22,18 @@ import net.minecraft.resources.Identifier;
 @Mixin(TitleScreen.class)
 public class TitleScreenMixin extends Screen {
 
-	private static final int ICON_ROW_BUTTON_WIDTH = 20;
-
 	protected TitleScreenMixin() {
 		super(null);
 	}
 
 	@Inject(method = "init", at = @At("TAIL"))
 	public void addWaylandCraftButton(CallbackInfo info) {
-		List<AbstractWidget> row = findIconRow();
+		List<AbstractWidget> row = IconRowUtils.findIconRow(this.children());
 
 		SpriteIconButton button = SpriteIconButton
 				.builder(Component.literal("waylandcraft"), (_) -> {Minecraft.getInstance().gui.setScreen(new WaylandCraftSettingsScreen(WaylandCraft.instance));}, true)
 				.sprite(Identifier.fromNamespaceAndPath(WaylandCraftCommon.MOD_ID, "logo"), 16, 16)
-				.width(ICON_ROW_BUTTON_WIDTH)
+				.width(IconRowUtils.ICON_ROW_BUTTON_WIDTH)
 				.build();
 
 		if(row.isEmpty()) {
@@ -47,58 +43,8 @@ public class TitleScreenMixin extends Screen {
 			return;
 		}
 
-		AbstractWidget rightmost = row.get(0);
-		for(AbstractWidget widget : row) {
-			if(widget.getX() > rightmost.getX()) rightmost = widget;
-		}
-
-		int spacing = computeSpacing(row);
-		int y = rightmost.getY();
-		int x = rightmost.getX() + rightmost.getWidth() + spacing;
-		button.setPosition(x, y);
+		IconRowUtils.joinRow(row, button);
 		this.addRenderableWidget(button);
-
-		int moveX = (button.getWidth() + spacing) / 2;
-		for(AbstractWidget widget : row) {
-			widget.setX(widget.getX() - moveX);
-		}
-		button.setX(button.getX() - moveX);
-	}
-
-	private static int computeSpacing(List<AbstractWidget> row) {
-		if(row.size() < 2) return 4;
-
-		int minX = Integer.MAX_VALUE;
-		int maxXEdge = Integer.MIN_VALUE;
-		int totalWidth = 0;
-		for(AbstractWidget widget : row) {
-			minX = Math.min(minX, widget.getX());
-			maxXEdge = Math.max(maxXEdge, widget.getX() + widget.getWidth());
-			totalWidth += widget.getWidth();
-		}
-
-		return Math.max(1, ((maxXEdge - minX) - totalWidth) / (row.size() - 1));
-	}
-
-	/**
-	 * Finds the largest group of same-width {@link SpriteIconButton}s sharing
-	 * a Y coordinate — the row of small square icon buttons vanilla and other
-	 * mods (e.g. accessibility, realms notifications, Mindful Darkness'
-	 * daytime toggle) place near the bottom-left of the title screen.
-	 */
-	private List<AbstractWidget> findIconRow() {
-		Map<Integer, List<AbstractWidget>> byY = new HashMap<>();
-		for(net.minecraft.client.gui.components.events.GuiEventListener listener : this.children()) {
-			if(!(listener instanceof SpriteIconButton widget)) continue;
-			if(widget.getWidth() != ICON_ROW_BUTTON_WIDTH) continue;
-			byY.computeIfAbsent(widget.getY(), (_) -> new ArrayList<>()).add(widget);
-		}
-
-		List<AbstractWidget> best = List.of();
-		for(List<AbstractWidget> group : byY.values()) {
-			if(group.size() > best.size()) best = group;
-		}
-		return best;
 	}
 
 }
